Finished some crude unit select display and troop mover. Started on the view clearance resolver. Looks promising so far, only needs testing now.

+++ Libhammer/Gamestate/BoardTroop.cs
@@ -13,7 +13,7 @@ public class BoardTroop
 
     // Movement stuffs
     public MoveStatus MoveState = MoveStatus.Pending;
-    public int AvailableMovement = 0;
+    public float AvailableMovement = 0;
 
     // Shooting stuffs
     public bool ShootingEligible = true;
